+++ attributes/kubelet.rb
@@ -23,6 +23,7 @@ default['kubernetes']['kubelet']['config']['clusterDomain']                     
 default['kubernetes']['kubelet']['config']['imageGCLowThresholdPercent']             = 70
 default['kubernetes']['kubelet']['config']['imageGCHighThresholdPercent']            = 80
 default['kubernetes']['kubelet']['config']['failSwapOn']                             = false
+default['kubernetes']['kubelet']['config']['cgroupDriver']                           = node['kubernetes']['cgroupdriver']
 default['kubernetes']['kubelet']['config']['readOnlyPort']                           = 10255
 default['kubernetes']['kubelet']['config']['serverTLSBootstrap']                     = true
 default['kubernetes']['kubelet']['config']['rotateCertificates']                     = true if node['kubernetes']['feature_gates']['RotateKubeletServerCertificate']