+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # UNRELEASED
 - `kubeletconfig`: check for `RotateKubeletServerCertificate` feature gate before enabling server certificates autorotation
 - Docker moved to separate recipe
+- install CRIO and corresponding binaries
 
 # 1.11.1 (28.06.2018)
 - Kubernetes: 1.11.0