+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - `kubeletconfig`: check for `RotateKubeletServerCertificate` feature gate before enabling server certificates autorotation
 - Docker moved to separate recipe
 - install CRIO and corresponding binaries
+- Explicitly install aufs-tools if cri-o uses aufs as storage
 
 # 1.11.1 (28.06.2018)
 - Kubernetes: 1.11.0